example tests are reorganized
