+++ examples/tsp/criteria/duration/mod.rs
@@ -1,5 +1,9 @@
+#[cfg(test)]
+mod tests;
+
 mod criterion;
 mod input;
 mod iter;
 mod move_generator;
-pub mod tests;
+
+pub use input::DurationMatrix;
