--- examples/tsp/criteria/duration/tests.rs
@@ -1,5 +1,0 @@
-use crate::criteria::duration::input::test_duration_matrix;
-
-pub fn run() {
-    test_duration_matrix();
-}
