+++ examples/tsp/criteria/duration/tests/mod.rs
@@ -1,0 +1,1 @@
+mod duration_matrix;
