+++ examples/tsp/main.rs
@@ -1,12 +1,18 @@
-mod neighborhood;
-mod criteria;
-mod insert_move;
-mod problem;
+#[cfg(test)]
 mod tests;
+
+pub mod criteria;
+mod insert_move;
+mod neighborhood;
+mod problem;
 mod tour;
 mod tour_after_move;
 mod tour_cost;
 
-fn main() {
-    tests::run();
-}
+pub use insert_move::InsertMove;
+pub use neighborhood::AllInsertMovesIter;
+pub use tour::Tour;
+pub use tour_after_move::TourAfterInsertIter;
+pub use tour_cost::TourCost;
+
+fn main() {}
