+++ examples/tsp/tests/mod.rs
@@ -1,0 +1,3 @@
+mod insert_move;
+mod insert_neighborhood_iter;
+mod tour_after_move;
